Refactor list membership hydration
